+++ NEWS.md
@@ -1,3 +1,9 @@
+## SynSigGen 1.1.0.9001
+### Fixed
+* Fixed a bug in function `AddNoise` when adding noise using
+Dirichlet-multinomial (some samples may not have exposures to a particular
+signature).
+
 ## SynSigGen 1.1.0.9000
 ### Added
 * Added new dependency package `MGLM` to allow drawing random numbers from